requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
Babel==2.8.0
bandit==1.6.2
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.0.8
docutils==0.16
gitdb==4.0.5
GitPython==3.1.7
gunicorn==20.0.4
idna==2.10
imagesize==1.2.0
importlib-metadata==1.7.0
isort==4.3.21
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.782
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
pbr==5.4.5
pycodestyle==2.6.0
pydocstyle==5.0.2
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
python-decouple==3.3
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
Sphinx==3.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
stevedore==3.2.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.10
whitenoise==5.1.0
wrapt==1.12.1
zipp==3.1.0
